+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 1.1.0
+
+- Add `Equatable` conformance to `LPSnackbar`
+- Add new parameter to `dismiss(show:)`, `completeWithAction: Bool` which allows you to dismiss the `LPSnackbar` manually while also returning `true` for action in the `SnackbarCompletion`.
+
 ## Version 1.0.0
 
 Initial release